Added check for empty value in setPower()
I created a pop up that prompts the user to enter in a value for setPower() if it is empty in the code. The user can enter the value and run the sim, but the code in the code tab won't be updated. If the user still doesn't enter a value or leave the prompt blank, the value is defaulted to 0.