Wheel Size Changes saved for Robot.txt files
I added the functionality for saving the wheel size to the robot.txt files so the wheel size can be preserved when saving a robot that is built. There is a small bug that causes the robot that is opened from a .txt file to load into the simulation tab a little higher on the grid than normal, but this can be fixed by clicking the "reset" or "clear" button. I will try to debug this later and patch it.