Line Sensor Part Additions (commented out)
I started working on the line sensor part, and added an HTML button for it and some of the back end code. There is an error occurring however, so I commented out the sections of the code. If I can't fix it by the end of the semester, I figured the code would still be helpful for future work.