HUD display change and drawing save functionality
I changed the HUD code to only appear when an opponent is present. I figured that the HUD is not necessary with only one robot and it would get in the way of drawings and such. I also add two new buttons to the simulate tab that allow the user to save the hand drawings and the robot drawings to PNG files.